requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.6.0
    # via pydantic
backoff==2.2.1
certifi==2024.2.2
    # via elastic-transport
elastic-transport==8.13.0
    # via elasticsearch
elasticsearch==8.13.0
psycopg2-binary==2.9.9
pydantic==2.7.1
    # via pydantic-settings
pydantic-core==2.18.2
    # via pydantic
pydantic-settings==2.2.1
python-dotenv==1.0.1
    # via pydantic-settings
typing-extensions==4.11.0
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via elastic-transport
